+++ tests/legacy/test_tron.py
@@ -13,6 +13,7 @@ import sys
 import pytest
 
 from clu.legacy import TronConnection, TronKey
+from clu.legacy.types.messages import Reply
 from clu.legacy.types.parser import ParseError
 
 
@@ -157,3 +158,15 @@ async def test_tron_connected(actor, tron_server):
     actor.tron.transport.close()
 
     assert actor.tron.connected() is False
+
+
+@pytest.mark.skipif(sys.version_info < (3, 8), reason="Test fails in PY37")
+async def test_reply_callback(actor, tron_server, mocker):
+
+    callback_mock = mocker.MagicMock()
+
+    command = actor.send_command("alerts", "ping", callback=callback_mock)
+    await command
+
+    callback_mock.assert_called()
+    assert isinstance(callback_mock.mock_calls[0].args[0], Reply)